Chat Channels Via Kafka (with Grading Notes)
In this assignment, we’ll take the work from the previous assignment on channelized chat and have it use
a distributed system as a backend instead of a custom server.  The distributed system that we’ll use is the
one that we saw in class recently:
Kafka.
The clients will speak a protocol built on the following message type:
senderName hour:minute:second - sentence
The line begins with the name of the sender followed by one
space (spaces aren’t allowed in the name). It’s followed by a time stamp in the format hour:minute:second
(
ex.
12:05:07) followed by a single space,  a dash (-),  and then single space.  After the second space
comes a sentence which may contain spaces.  The message ends with a new line character.
The senderName is the sender name as chosen by the sender.  It is not managed by Kafka and does
not need to be globally unique.
The timestamp is the local time at the sender’s computer.
Messages of the above type are prepared and sent by Kafka
Producer
objects for distribution to clients
who have
Consumer
objects for the message’s topic.
Note that the message format is different from the previous assignment. Since Kafka will forward messages
based  on  topics  and  handle  all  topic  registration  and  leaving,  there  is  no  need  for  other  message  types.
Depending on how you build your clients, you may not need to parse the above messages at all since Kafka
will distribute the messages between the
Consumer
objects automatically.